+++ app/views/Application/index.html
@@ -10,7 +10,7 @@
 		<a class="btn primary large" href="@{Login.index()}">Se connecter</a>
 		#{/if}
 		#{else}  
-		<a class="btn primary large" href="@{Application.showMember(session.get("username"))}">Mon profil</a>
+		<a class="btn primary large" href="@{Profile.show(session.get("username"))}">Mon profil</a>
 		#{/else}  
 	</div>
 </div><!-- /container -->
