+++ app/views/Application/list.html
@@ -25,7 +25,7 @@
 			  <tr>
 				  <td>${member.firstname}</td> 
     	          <td>${member.lastname}</td>
-    	          <td><a href="@{Application.showMember(member.login)}">Profil</a></td>
+    	          <td><a href="@{Profile.show(member.login)}">Profil</a></td>
     	      </tr>
 		#{/list}
 		</tbody>
